requirements.txt
# FastAPI and web framework dependencies
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Template engine
jinja2==3.1.3

# Database drivers
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Redis client
redis==5.0.1

# Image processing
Pillow==10.2.0

# AI/ML libraries
transformers==4.37.2
torch==2.2.0

# Additional utilities
python-dotenv==1.0.0

# Testing
pytest==7.4.3
hypothesis==6.92.1
